requirements.txt
psycopg2-binary
python-dotenv
folium
shapely
requests
ultralytics
pyproj
dash
plotly